requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
streamlit
python-dotenv
groq
